+++ src/support/hash.h
@@ -2,4 +2,4 @@
 
 #include <redasm/config.h>
 
-u32 rd_i_murmur3(const char* s, usize len);
+u32 rd_i_murmur3(const char* s, u32 len);
